The semantic actions at the end, i.e., to delete the auxiliary label zLabelToStopActions, were eliminated. They cause some weird label name in the filters action, as we could see in gmail.
The presence of the label zLabelToStopActions is not really a
problem, as it has an mostly a comestic effect at the label list; so it
will remain at the label list.

But it could a great comestic effect if we could eliminate this label
from all messages, instead of eliminate the label itself from the
label list.